# ENH: optimize std in scipy.stats.binned_statistic