<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
https://docs.scipy.org/doc/numpy/dev/development_workflow.html#writing-the-commit-message

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### What does this implement/fix?
The current version of `scipy.stats.binned_statistic_dd` allows the following values for the statistic parameter:
['mean', 'median', 'count', 'sum', 'std', 'min', 'max'] (in addition to any callable)
Of these, sum, count, and mean are computed in a vectorized way while the rest are passed in a loop to the corresponding numpy functions which leads to much slower performance.

This PR aims to vectorize the `std` statistic calculation since it actually only requires knowledge of the `count`, and `mean` results (which as noted were already being done in a vectorized way).

Previously, I already addressed the `min`, `max`, and `median` in https://github.com/scipy/scipy/pull/14625 and just realized that `std` can be optimized as well. Taken together, these PR's can enable all of the "named" statistics to be calculated in a vectorized way with much better benchmark performance.

#### Additional information
Note: This also affects `binned_statistic` and `binned_statistic_2d`

On my machine, the benchmarks *before* this change:

            count                   574±4μs   
             sum                   556±0.2μs  
             mean                  618±0.2μs  
             min                  11.9±0.09ms 
             max                  12.3±0.08ms 
            median                 68.9±0.1ms 
             std                  4.12±0.01ms 
             <function std at 0x7f0d86328940>    47.1±0.3ms 

and *after*: (see `std`)

            count                   573±6μs   
             sum                   557±0.4μs  
             mean                   616±2μs   
             min                   11.8±0.1ms 
             max                  11.7±0.05ms 
            median                 67.4±0.3ms 
             std                   656±100μs  
             <function std at 0x7fbf29174940>    46.6±0.2ms