@rgommers 2021-09-10
I have added the release note at https://github.com/scipy/scipy/wiki/Release-note-entries-for-SciPy-1.8.0